refactor[data_storage]: split integration tests into modules
